Retrieve WorkingDirectory differently
This solution doesn't suffer from cross-platform issues. The previous
way was broken on Linux, because the "file:///" pattern was removing the
initial slash of a Linux absolute path, thus making it relative and
breaking everything. However, on Windows removing that initial slash is
actually necessary to get a valid path.