Add workaround for joystick OverflowException
I have no idea how an OverflowException is possible as shown on the
backtraces from the players. The line where it occurs has no arithmetic
operations or casts. It makes no sense whatsoever.

However, plenty of players hit the issue, so this commit adds a
workaround. If an OverflowException is encountered, the code prints a
warning and stops updating the joystick state.